requirements.txt
Django==1.8.2
Pillow==2.8.1
django-taggit==0.14.0
-e git+git://github.com/zenx/django-thumbs.git@develop#egg=django-thumbs